+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
-- Support ActiveSupport::Cache.lookup\_store
+- Support `ActiveSupport::Cache.lookup_store`
+- `#write` and `#delete` now return responses from underlying caches
 
 ## [0.1.0] - 2024-02-13